streamlit/requirements.txt
streamlit
pvporcupine
pyaudio
openai
pyttsx3
requests
beautifulsoup4
python-dotenv
pyyaml
sounddevice
webrtcvad
openai-whisper
psutil
fastapi
uvicorn[standard]
notion-client
pytest
pytest-mock
pytest-asyncio
pytest-cov
httpx